Implemented remote document, vertex and edge classes
1st phase of issue #497.

Now the right object type is returned. It's not possible to have a complete 100% abstraction of the embedded database, but we're much closer. Now you can operate with vertices and edges.

In the next phase we will allow to `modify()` document, vertex and edges by using a remote abstraction of mutable records.

And maybe  in the long run we could let RemoteDatabase class implement Database class and provide a full or almost full abstraction.