fix: rename HTTP graph serialization
After many thoughts, I'd prefer to change "graph" to serializing only the graph element. This is because everybody is using "graph" but nobody is interested in the "record" collections, and it's a waste of RAM and bandwidth.

So I'm going to rename "strictgraph" to just "graph" and the old "graph" in "studio", because it's the way Studio mainly interacts to get both graph and record elements. The graph is to draw the graph, records to fill the table.

Issue #673