Speedup query execution by removing not necessary operations
While digging into the query execution with the profiler, there were a lot of calls to the toString() of the element of the query AST. This PR optimizes the usage of toString() by caching the string representation.

Also, context variables are now always set without the dollar sign ($) as prefix saving one copy of the string per execution.