fix: query with rid as projection and order by asc returned an internal column
Fixed issue #725

There is a part in the SQL grammar that never supported ORDER BY @rid. Just checked the code for the OrientDB SQL layer and it's the same. I guess the developer started something but ended up using @rid as a record attribute instead of a proper identifier.

The reason why the internal alias is set in the result is that the engine doesn't recognize you used @rid in your projection, so it creates an additional property to index the RID.