Removed unused `Resultset.countEntries()` to avoid confusion
Fixed issue #728

The SQL engine tries as much as it's possible, to use RAM, so a lot of operations are lazy evaluated. This means it's not possible to know the resultset in many cases unless the whole result set is iterated that means the whole query is executed.

There is no clean way to fix this. Since the API is only used in tests and it's very misleading (even with a bold javadoc), I'd prefer to remove it from the code base and move it as `CollectionUtils.countEntries(iterator)`.