fix: got rid of org.json library and used GSON instead
Fixed issue #726. The issue was with the Java library org.json that does not support ordered properties. I sent a PR to the project more 1y ago, and after 1y they decide to not include it... So we're not using that piece of sw anymore. In order to make this transition easier, I've implemented the same JSONObject, JSONArray and JSONException of the previous JSON parser but under the hood uses GSON. This also make things easier in case we need to migrate to something else in the future.