fix:range query with against compacted indexes
Issue #757

After a long debug session (the compacted LSM-Tree index is the most complex component of ArcadeDB) I decided to write a new LSM-Tree chapter in the docs, under Appendix, for whom is interested in contributing to LSM-Tree.

The issue was in the range query: after compaction, the entries are sparse in multiple segments. The range() operation didn't check for the upper bound of the keys, so some of the pages containing the values were discarded when the iterators were built.