perf: refactoring of the page manager component to completely remove the write cache
The write cache added overhead without any real benefit. Now the page is cached as immutable in the read cache as soon as the page manager writes the page in the log.

The modified page is sent to the flush thread that will asynchronously write to disk (unless async=false).

The consistency is assured by the WAL, so even if the page is not written to disk, at the next restart it will be written from the log.