perf: improved performance of comparison between not compatible types and nulls
This fix dramatically improves the performance of traversing when different types are encountered or simply null values. This piece of code should be designed to be fast because invoked millions of times with queries. It is a bad practice to throw Java exceptions instead of simply returning false.

Open a PR for this: https://github.com/apache/tinkerpop/pull/1993