fix: missed backup database default directory
Fixed issue #1069

The problem is with the setting of the global server root path. In case of multiple servers on the same JVM (usually with replication tests only), the setting is saved in the local server configuration and not in the global. The SQL command didn't have a way to access to the server configuration because independent.

I've added a new method signature to provide the configuration so all the commands executed through the server can now pass the server cfg.