fix: fixed bug with management of TX isolation from cached vertices
The solution is in the force of reloading of the record when the page is not in transaction. This assure the record is the same as for the loaded page. Without this fix, the record can be outdated managed with the loading of an updated page, so the MVCC check passes causing a overwrite of a new record with an old one.