Releasing ArcadeDB 23.12.2 because of issues with sonatype/maven central
The latest release, 23.12.1, wasn't published to Maven Central because there was an outage at that time, then resolved by Sonatype in a few hours. After that, seems like Sonatype has mixed up the jars and doesn't allow us to proceed with the release, because it's registered as already release, but there are no jars online... I'm gonna try to release again as 23.12.2 that is identical to 23.12.1, hoping with a fresh release workflow everything works.