feat: implemented space reclaim from pages v1
This v1 algorithm keeps a list of max 20 pages with statistics of number of records and free space per page. When new space is needed (insert/update), then the list of pages is checked first to recycle space from old pages. The statistics are kept in a json file written at db closing, like the count file.

This file is not crucial, can be lost, it's just an approx. cache of free space available.